# gh-104066: Improve performance of hasattr for module objects