--- a/Misc/NEWS.d/next/Core and Builtins/2023-05-01-14-48-29.gh-issue-104066.pzoUZQ.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2023-05-01-14-48-29.gh-issue-104066.pzoUZQ.rst
@@ -1,0 +1,2 @@
+Improve the performance of :func:`hasattr` for module objects with a missing
+attribute.